Add entire .idea folder to .gitignore file
This folder is used for project configuration within Android
Studio, but contains OS-specific information and therefore
should not be commited into version control. Gradle will handle
the population of the items necessary for the project into
these folders.

This also makes it easier if you later want to move this project
to another IDE.