Add an example of a custom view to the application
This goes through and adds a custom view to the app, currently
showing on the PitActivity. This view shows how you can package
a custom UI element that is defined in an XML layout, define
custom attributes for that XML layout, and process information
within that view.

This makes it easier to incorporate the same UI elements on
multiple screens, packaging all of the logic needed for that
UI element in a single place.